model/requirements.txt
ipykernel
numpy
pandas
scikit-learn
scipy
dotenv
google-generativeai